pyproject.toml
[tool.poetry]
name = "keepyrs"
version = "0.1.0"
description = ""
authors = ["frank <frank@lunoho.company>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anchorpy = "0.17.1"
solana = "0.30.1"
python-dotenv = "^1.0.0"
drift-jit-proxy = "^0.1.1"
driftpy = {path = "/Users/frankfinnegan/Documents/musings/drift/driftpy", develop = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"